fix(gui): selectedVideoPath を保存前に \?\ prefix 正規化 (Refs #466)
[elated-wozniak-50d3bb]

実機テスト 4 で発見した \"video file not found: \?\E:\videos\..\" エラー対策。
原因: Tauri dialog/drag-drop が Windows で返す extended-length path
prefix が selectedVideoPath にそのまま入り、Rust validate_video_path で
Path::is_file() が false を返すケースがあった。

修正:
- utils/path.ts に stripExtendedPathPrefix を新規作成 (ExportScreen から
  移動し、共通 helper 化)
- appStateStore.setSelectedVideoPath で正規化を集中させる。これで
  selectedVideoPath を読む全 consumer (ExportScreen の videoSource、
  PreviewScreen の register_video / generate_match_thumbnails、UI 表示)
  が一貫して prefix なしの path を扱う
- deriveDefaultOutDir でも defense-in-depth として strip 継続

テスト:
- utils/path.test.ts: drive-letter / UNC / pass-through 3 ケース
- appStateStore: setSelectedVideoPath が \?\ / \?\UNC / pass-through を
  正しく扱う 3 ケース追加
- ExportScreen から重複していた stripExtendedPathPrefix 単体テストは削除

vitest: 29 files / 323 passed (前回 320 → +3)
eslint / tsc --noEmit: clean
